refactor(network): move network utils to cli-framework
